[CustomForm] Load JS/css files from resource files
