+++ src/models/custom-form-resources/form.css
@@ -1,0 +1,171 @@
+.formSection {
+    border: 1px solid #cacaca;
+    border-radius: 3px;
+    margin: 0;
+    padding: 1px 4px 1px 4px;
+}
+
+.formSection h3 {
+    color: #000;
+    font-size: 16px;
+    text-align: left;
+    font-weight: bold;
+    padding: 0;
+}
+
+.formSection td {
+    text-align: center;
+    min-width: 75px;
+    padding: 1;
+}
+
+.formSection th {
+    font-weight: normal;
+    font-size: 14px;
+    text-align: center;
+    white-space: normal !important;
+    max-width: 75px;
+    background-color: #eaf7fb;
+    max-width: 200px;
+}
+
+.entryfield,
+.total-cell,
+.indicator {
+    max-width: 75px;
+}
+
+.entryarea {
+    max-width: 300px;
+}
+
+#contentDiv input.entryfield {
+    width: 70px;
+    height: 18px;
+    padding: 2px;
+}
+
+#contentDiv input.dataelementtotal {
+    width: 70px;
+    height: 16px;
+    padding: 2px;
+}
+
+#contentDiv input.indicator {
+    width: 70px;
+    height: 18px;
+    padding: 2px;
+}
+
+#contentDiv tr {
+    border-color: transparent;
+    transition: background-color 500ms linear;
+}
+
+#contentDiv tr.derow {
+    background-color: #fff;
+}
+
+#contentDiv tr.hover,
+#contentDiv tr:hover {
+    background-color: #e5e5e5;
+}
+
+#contentDiv tr.focus {
+    background-color: #e5f5e5;
+}
+
+#contentDiv th {
+    border-style: hidden !important;
+}
+
+#contentDiv td {
+    padding: 2px !important;
+    height: 16px;
+    text-align: center;
+    border-style: none !important;
+}
+
+#contentDiv td.data-element {
+    text-align: left;
+    white-space: normal;
+    max-width: 33%;
+}
+
+#contentDiv tr.secondary td.data-element {
+    font-style: italic;
+    opacity: 0;
+}
+
+#contentDiv .header-first-column {
+    background-color: #e0e0e0;
+    text-align: left;
+    border-bottom-style: hidden;
+    border-left-style: hidden;
+    border-top-style: hidden;
+    white-space: nowrap;
+    background-color: #fff;
+    padding: 2px !important;
+}
+
+#contentDiv .header-first-column.full-width {
+    width: 100%;
+}
+
+#contentDiv th.data-header {
+    text-align: center;
+    border-bottom: 1px solid #ddd !important;
+    background-color: #eaf7fb;
+    white-space: nowrap;
+    padding-top: 2px !important;
+    padding-bottom: 2px !important;
+    padding-left: 5px;
+    padding-right: 5px;
+    word-wrap: break-word;
+}
+
+#contentDiv .panel-default > .panel-heading {
+    background-color: #3c3c3c !important;
+    border-color: #3c3c3c;
+    padding: 7px;
+    margin-bottom: 5px;
+    cursor: pointer;
+}
+
+.ui-state-active a,
+.ui-state-active a:link,
+.ui-state-active a:visited {
+    color: #053e21;
+}
+
+.ui-widget-header {
+    background: none;
+    background-color: #2f3867;
+}
+
+#contentDiv .dataValuesTable {
+    margin-bottom: 0px !important;
+    margin-top: 5px;
+}
+
+.page th {
+    text-align: left;
+    color: #39547d;
+    padding: 3px 0 3px 1px;
+    font-size: 13px;
+    font-weight: bold;
+    border-collapse: collapse;
+    border-bottom: 1px solid #cad5e5;
+    min-height: 28px;
+}
+
+.dataelement-group {
+    margin-bottom: 20px;
+}
+
+.dataelement-group .title {
+    color: #544;
+    font-size: 1.2em;
+    font-weight: bold;
+    margin: 5px 0px 5px 0px;
+}
